+++ bchannel.h
@@ -13,7 +13,7 @@
 struct bchannel {
 	struct bchannel *next;
 	struct chan_call *call;		/* link to call process */
-	unsigned long handle;		/* handle for stack id */
+	unsigned int handle;		/* handle for stack id */
 	int b_sock;			/* socket for b-channel */
 	int b_mode;			/* dsp, raw, dsphdlc */
 	int b_state;
@@ -22,7 +22,7 @@ struct bchannel {
 	int b_tx_dejitter;
 	int b_tx_gain, b_rx_gain;
 	char b_pipeline[256];
-	unsigned long b_conf;
+	unsigned int b_conf;
 	int b_echo;
 	int b_tone;
 	int b_rxoff;
@@ -48,8 +48,8 @@ void bchannel_blowfish(struct bchannel *bchannel, unsigned char *key, int len);
 void bchannel_pipeline(struct bchannel *bchannel, char *pipeline);
 void bchannel_gain(struct bchannel *bchannel, int gain, int tx);
 int bchannel_handle(void);
-struct bchannel *find_bchannel_handle(unsigned long handle);
-//struct bchannel *find_bchannel_ref(unsigned long ref);
-struct bchannel *alloc_bchannel(unsigned long handle);
+struct bchannel *find_bchannel_handle(unsigned int handle);
+//struct bchannel *find_bchannel_ref(unsigned int ref);
+struct bchannel *alloc_bchannel(unsigned int handle);
 void free_bchannel(struct bchannel *channel);
